requirements.txt
matplotlib==3.6.2
numdifftools==0.9.41
numpy==1.23.4
PyYAML==6.0
